Prepare for Admin Enhance security for user me

+++ routers/routers_user_me.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
 from sqlalchemy.orm import Session
 
 import crud
@@ -28,18 +28,34 @@ def read_user_me(current_user: schemas.UserOut = Depends(get_current_user)):
     status_code=status.HTTP_200_OK
 )
 def update_user_me(
-        user: schemas.UserUpdate,
+        user: schemas.UserUpdateWithPassword,
         db: Session = Depends(get_db),
         current_user: schemas.UserOut = Depends(get_current_user)):
-    return crud.update_user(db, current_user.id, user)
+    if crud.authenticate_user(db, current_user.phone_number, user.password):
+        return crud.update_user(db, current_user.id, user)
+
+    raise HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Incorrect password"
+    )
 
 
 @router.delete(
     '/',
     status_code=status.HTTP_204_NO_CONTENT,
 )
-def delete_user_me(db: Session = Depends(get_db), current_user: schemas.UserOut = Depends(get_current_user)):
-    crud.delete_user(db, current_user.id)
+def delete_user_me(
+        user: schemas.UserDeleteWithPassword,
+        db: Session = Depends(get_db),
+        current_user: schemas.UserOut = Depends(get_current_user)
+):
+    if crud.authenticate_user(db, current_user.phone_number, user.password):
+        crud.delete_user(db, current_user.id)
+
+    raise HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Incorrect password"
+    )
 
 
 @router.get(
